requirements.txt
python-dotenv==1.0.0
SQLAlchemy==2.0.25
python-telegram-bot==20.7
greenlet==3.1.1
cryptography==44.0.1
aiomysql==0.3.0
